feat: 누락 테스트 추가 및 추가된 API RestDocs 적용 (#52) (#53)
* style: log에 '' 제거

* refactor: 모든 api에 대한 test 추가, 문서 생성
