+++ backend/src/main/resources/application-prod.properties
@@ -10,8 +10,8 @@ logging.level.org.springframework.web= debug
 logging.level.sql= error
 logging.level.jdbc= error
 logging.level.jdbc.sqlonly= debug
-logging.pattern.console= '%-5level %d{yyyy-MM-dd HH:mm:ss}[%thread] [%F - %M] [%logger{0}:%line] - %msg%n'
-logging.pattern.file= '%-5level %d{yyyy-MM-dd HH:mm:ss}[%thread] [%F - %M] [%logger{0}:%line] - %msg%n'
+logging.pattern.console= %-5level %d{yyyy-MM-dd HH:mm:ss}[%thread] [%F - %M] [%logger{0}:%line] - %msg%n
+logging.pattern.file= %-5level %d{yyyy-MM-dd HH:mm:ss}[%thread] [%F - %M] [%logger{0}:%line] - %msg%n
 
 logging.file.name= ${user.home}/logs/zzimkkong.log
 logging.logback.rollingpolicy.max-file-size= 100MB
