+++ backend/src/test/java/com/woowacourse/zzimkkong/controller/MemberControllerTest.java
@@ -2,39 +2,21 @@ package com.woowacourse.zzimkkong.controller;
 
 import com.woowacourse.zzimkkong.dto.MemberSaveRequest;
 import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import static com.woowacourse.zzimkkong.controller.DocumentUtils.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
-@ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
-@AutoConfigureRestDocs
 class MemberControllerTest extends AcceptanceTest {
     public static final String EMAIL = "pobi@email.com";
     public static final String PASSWORD = "test1234";
     public static final String ORGANIZATION = "루터";
-
-    @BeforeEach
-    public void setUp(RestDocumentationContextProvider restDocumentation) {
-        RequestSpecification spec = new RequestSpecBuilder()
-                .addFilter(documentationConfiguration(restDocumentation))
-                .build();
-        DocumentUtils.setRequestSpecification(spec);
-    }
 
     @DisplayName("정상적인 회원가입 입력이 들어오면 회원 정보를 저장한다")
     @Test
@@ -59,16 +41,29 @@ class MemberControllerTest extends AcceptanceTest {
         saveMember(memberSaveRequest);
 
         //then
-        ExtractableResponse<Response> response = validateDuplicateEmail();
+        ExtractableResponse<Response> response = validateDuplicateEmail(EMAIL);
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    private ExtractableResponse<Response> validateDuplicateEmail() {
+    @DisplayName("중복되지 않은 이메일을 입력하면 200 ok를 반환한다.")
+    @Test
+    void getMembers() {
+        //given
+        String email = "dusdn1702@naver.com";
+
+        //when
+        ExtractableResponse<Response> response = validateDuplicateEmail(email);
+
+        //then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    private ExtractableResponse<Response> validateDuplicateEmail(String email) {
         return RestAssured
-                .given(DocumentUtils.getRequestSpecification()).log().all()
+                .given(getRequestSpecification()).log().all()
                 .accept("application/json")
-                .filter(document("member", DocumentUtils.getRequestPreprocessor(), DocumentUtils.getResponsePreprocessor()))
-                .queryParam("email", EMAIL)
+                .filter(document("member/get", getRequestPreprocessor(), getResponsePreprocessor()))
+                .queryParam("email", email)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/api/members")
                 .then().log().all().extract();
